Вынесены в отдельные файлы модели из MovieQuizViewController: генерация вопросов, структуры.
