+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,14 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Artem Dubovitsky on 27.02.2023.
+//
+
+import Foundation
+
+struct QuizQuestion {
+    let image: String
+    let text = "Рейтинг этого фильма больше чем 6?"
+    let correctAnswer: Bool
+}
