+++ MovieQuiz/Services/QuestionFactory.swift
@@ -1,0 +1,51 @@
+//
+//  QuestionFactory.swift
+//  MovieQuiz
+//
+//  Created by Artem Dubovitsky on 27.02.2023.
+//
+
+import Foundation
+
+class QuestionFactory: QuestionFactoryProtocol {
+    private let questions: [QuizQuestion] = [
+        QuizQuestion(
+            image: "The Godfather",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "The Dark Knight",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "Kill Bill",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "The Avengers",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "Deadpool",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "The Green Knight",
+            correctAnswer: true),
+        QuizQuestion(
+            image: "Old",
+            correctAnswer: false),
+        QuizQuestion(
+            image: "The Ice Age Adventures of Buck Wild",
+            correctAnswer: false),
+        QuizQuestion(
+            image: "Tesla",
+            correctAnswer: false),
+        QuizQuestion(
+            image: "Vivarium",
+            correctAnswer: false)
+    ]
+}
+
+func requestNextQuestion() -> QuizQuestion? {
+    guard let index = (0..<questions.count).randomElement() else {
+        return nil
+    }
+    return questions[safe: index]
+}
+
